Separate flattening and expanding functions to different file in paired router to GCP connection resource
